move component into /page/posts/
