+++ src/pages/index.tsx
@@ -3,8 +3,8 @@ import { InferGetStaticPropsType, GetStaticProps } from "next"
 import Head from "next/head"
 import { Layout } from "components/Layout/Layout"
 import { Wrapper } from "components/Wrapper/Wrapper"
-import { ArticleCard } from "components/ArticleCard/ArticleCard"
-import { ArticleCardWrapper } from "components/ArticleCard/ArticleCardWrapper"
+import { ArticleCard } from "components/page/posts/ArticleCard/ArticleCard"
+import { ArticleCardWrapper } from "components/page/posts/ArticleCardWrapper/ArticleCardWrapper"
 import { PostData } from "../components/interface/post-data"
 import { BlogPostRepository } from "../infra/blog-post-repository"
